BRG-2219: Add workaround for incorrect ListView behavior on Android
ListView always has unclamped scrolling on Android (but not iOS) - also present when using the Simulator with an Android / iOS device preset. Seems to be a bug in UI Toolkit, for now the workaround is to check for indexes outside of the range of the items, and ignore those Bind calls. Lists will still scroll incorrectly though.